backend/requirements.txt
# Versions reviewed and fixed on 2025-07-02
fastapi==0.111.0
uvicorn==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.31
python-dotenv==1.0.1
pytest==8.2.2
httpx==0.27.0
fastapi-limiter==3.2.0
redis==5.0.7
google-auth-library==2.30.0
pyotp==2.9.0
qrcode==7.4.2
fastapi-csrf-protect==1.2.0
structlog==24.2.0
pydantic-settings==2.3.4